Generalized Code for Arbitraty Resonator States
This is it right here chief. Turns out that in general, phi0 is indeed invariant regardless of modulation of a resonator. We can use the slopes calculated from this to develop the simple starting firmware for basic tone tracking. Also includes s2p files if you want to run it on your own.